requirements.txt
fastapi==0.115.6
uvicorn==0.19.0
python-dotenv==1.1.0
python-multipart==0.0.6
requests==2.32.3
openai==1.13.3
twilio==9.1.0
airtable-python-wrapper==0.15.3